pyproject.toml
[project]
name = "mle-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain-google-genai>=2.1.5",
    "langchain-google-vertexai>=2.0.26",
    "langchain[anthropic]>=0.3.26",
    "langgraph>=0.4.10",
    "python-dotenv>=1.1.1",
]
